Create TaskManager to have method about Task that be moved method from class
